// wwg-backend/src/type/responseMessages.type.ts
export interface ResponseMessagesType{
    todoAddPositive: string
    todoAddNegative: string
    todoDeletePositive: string
    todoDeleteNegative: string
    todoUpdatePositive: string
    todoUpdateNegative: string
    todoGetPositive: string
    todoGetNegative: string
    todoGetAllPositive: string
    todoGetAllNegative: string
    badRequest: string,
    internalServerError: string
}